Added htcondor run cancel

+++ qanat/cli_commands/run.py
@@ -87,36 +87,42 @@ def delete_run(experiment_name: str, run_id: int):
                 run.runner)
 
             # Cancel the run
-            info = execution_handler(Session, run.id).parse_yaml_file()
+            if run.runner == "local":
+                # Since the main process is somewhere else, we need to
+                # do stuff here.
+                # TODO: try to integrate in execution_handler
+                info = execution_handler(Session, run.id).parse_yaml_file()
 
-            # Kill main pid
-            try:
-                os.kill(info['main_pid'], signal.SIGTERM)
-                wait_finish = True
-            except ProcessLookupError:
-                logger.debug(f"Process {info['main_pid']} not found")
+                # Kill main pid
+                try:
+                    os.kill(info['main_pid'], signal.SIGTERM)
+                    wait_finish = True
+                except ProcessLookupError:
+                    logger.debug(f"Process {info['main_pid']} not found")
 
-            logger.info(
-                    f"Run {run_id} of experiment {experiment_name} canceled")
-
-            # Wait for signal that run has been canceled
-            if wait_finish:
-                session.close()
-                console = rich.console.Console()
-                with console.status(
-                        "[bold green]Waiting for run to finish gracefully..."):
-                    cancel_done = False
-                    while not cancel_done:
-                        session = Session()
-                        run = session.query(RunOfAnExperiment).filter_by(
-                            experiment_id=experiment_id, id=run_id).first()
-                        time.sleep(0.5)
-                        cancel_done = run.status == "cancelled"
-                        session.close()
-                # Close the database if not closed
-                if session.is_active:
+                # Wait for signal that run has been canceled
+                if wait_finish:
                     session.close()
-                session = Session()
+                    console = rich.console.Console()
+                    with console.status(
+                            "[bold green]Waiting for run to finish gracefully..."):
+                        cancel_done = False
+                        while not cancel_done:
+                            session = Session()
+                            run = session.query(RunOfAnExperiment).filter_by(
+                                experiment_id=experiment_id, id=run_id).first()
+                            time.sleep(0.5)
+                            cancel_done = run.status == "cancelled"
+                            session.close()
+                    # Close the database if not closed
+                    if session.is_active:
+                        session.close()
+                    session = Session()
+            
+            else:
+                execution_handler(Session, run.id).cancel_experiment()
+                logger.info(
+                        f"Run {run_id} of experiment {experiment_name} canceled")
 
         delete_run_from_id(session, run_id)
         logger.info(f"Run {run_id} of experiment {experiment_name} deleted")
